chore: refactor tests
